chore: move re* plugins to folder
