+++ src/config/mdx.ts
@@ -19,15 +19,13 @@ import remarkMdxFrontmatter from "remark-mdx-frontmatter";
 import { convertCompilerOptionsFromJson } from "typescript";
 
 import type { SolidBaseResolvedConfig } from "./index.js";
-import { rehypeFixExpressiveCodeJsx } from "./rehype-plugins.js";
-import {
-	remarkAddClass,
-	remarkDirectiveContainers,
-	remarkGithubAlertsToDirectives,
-	remarkIssueAutolink,
-	remarkRelativeImports,
-	remarkTOC,
-} from "./remark-plugins.js";
+import { rehypeFixExpressiveCodeJsx } from "./rehype-plugins/fix-expressive-code.js";
+import { remarkDirectiveContainers } from "./remark-plugins/directives";
+import { remarkGithubAlertsToDirectives } from "./remark-plugins/gh-directives";
+import { remarkIssueAutolink } from "./remark-plugins/issue-autolink";
+import { remarkAddClass } from "./remark-plugins/kbd";
+import { remarkRelativeImports } from "./remark-plugins/realtive-imports";
+import { remarkTOC } from "./remark-plugins/toc";
 
 export function solidBaseMdx(sbConfig: SolidBaseResolvedConfig<any>) {
 	return mdx.default.withImports({})({
